refactor: tailwind-class 파일명 변경
