+++ frontend/components.json
@@ -13,6 +13,6 @@
     "aliases": {
         "components": "@/components",
         "ui": "@/components/common",
-        "utils": "@/utils/tailwind-class"
+        "utils": "@/utils/tailwindClass"
     }
 }
